+++ vue/src/lib/select/examples/index.js
@@ -1,0 +1,11 @@
+import Select from '../index.vue';
+import ExampleSelectDefault from './default.vue';
+import ExampleSelectMultiSelect from './multi-select.vue';
+
+
+export default {
+  Select,
+  ExampleSelectDefault,
+  ExampleSelectMultiSelect
+};
+