+++ vue/CHANGELOG.md
@@ -3,6 +3,25 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.0.0-beta.26](https://github.com/momentum-design/momentum-ui/compare/@momentum-ui/vue@1.0.0-beta.25...@momentum-ui/vue@1.0.0-beta.26) (2019-10-26)
+
+
+### Features
+
+* **Vue:** add Select ([4ee4c2a](https://github.com/momentum-design/momentum-ui/commit/4ee4c2a))
+
+
+
+
+
+# [1.0.0-beta.25](https://github.com/momentum-design/momentum-ui/compare/@momentum-ui/vue@1.0.0-beta.24...@momentum-ui/vue@1.0.0-beta.25) (2019-10-25)
+
+**Note:** Version bump only for package @momentum-ui/vue
+
+
+
+
+
 # [1.0.0-beta.24](https://github.com/momentum-design/momentum-ui/compare/@momentum-ui/vue@1.0.0-beta.23...@momentum-ui/vue@1.0.0-beta.24) (2019-10-24)
 
 **Note:** Version bump only for package @momentum-ui/vue